+++ src/pygrape/compressed_delaf.py
@@ -1,6 +1,5 @@
-from pygrapenlp.pygrape import u_read_compressed_delaf, compressed_delaf
-from pygrapenlp.u_array import string_to_u_array
-from pygrapenlp.u_set_trie import add_u_set_trie_strings_to_string_set
+from pygrape import compressed_delaf, u_read_compressed_delaf, string_to_u_array, u_set_trie_to_string_set, \
+    add_u_set_trie_strings_to_string_set
 
 
 class CompressedDelaf:
