+++ src/pygrape/u_out_bound_trie.py
@@ -1,5 +1,5 @@
-from pygrapenlp.pygrape import u_array
-from pygrapenlp.u_array import u_array_to_string
+from pygrape import u_array_to_string
+from pygrape.pygrape import u_array
 
 
 def u_out_bound_trie_string_to_string(native_u_out_bound_trie_string):
