+++ test/unit/test_pygrapenlp.py
@@ -4,8 +4,8 @@ from collections import OrderedDict
 
 import pytest
 
-from pygrapenlp.grammar_engine import GrammarEngine
-from pygrapenlp.u_out_bound_trie import u_out_bound_trie_string_to_string
+from pygrape.grammar_engine import GrammarEngine
+from pygrape.u_out_bound_trie import u_out_bound_trie_string_to_string
 
 tag_test_cases = [
     ('', '{}'),
