+++ test/unit/u_array_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pygrapenlp.u_array import string_to_u_array, u_array_to_string
+from pygrape.u_array import string_to_u_array, u_array_to_string
 
 encode_string_decode_u_array_test_cases = [
     (''),
